Fix slow build tag compilation
