+++ compiler/meta/meta.go
@@ -1,3 +1,5 @@
+//go:build slow
+
 package meta
 
 import (
