+++ compiler/x/cpp/helpers.go
@@ -1,3 +1,5 @@
+//go:build slow
+
 package cpp
 
 import (
