+++ compiler/x/erlang/tools.go
@@ -1,3 +1,5 @@
+//go:build slow
+
 package erlang
 
 import (
