+++ compiler/x/st/rosetta_golden_test.go
@@ -1,3 +1,5 @@
+//go:build slow
+
 package st_test
 
 import (
