+++ compiler/x/zig/job_golden_test.go
@@ -17,13 +17,6 @@ import (
 	"mochi/types"
 )
 
-func stripHeader(b []byte) []byte {
-	if i := bytes.IndexByte(b, '\n'); i != -1 && bytes.HasPrefix(b, []byte("// Generated by Mochi")) {
-		return b[i+1:]
-	}
-	return b
-}
-
 // TestZigCompiler_JOB_Golden compiles JOB query q1 and compares the
 // generated Zig code with the golden file. The compiled executable is
 // then run and its output compared with the VM.
