+++ compiler/x/zig/testhelpers_test.go
@@ -1,0 +1,22 @@
+//go:build slow
+
+package zigcode_test
+
+import (
+	"bytes"
+	"flag"
+)
+
+var update = flag.Bool("update", false, "update golden files")
+
+func shouldUpdate() bool {
+	f := flag.Lookup("update")
+	return f != nil && f.Value.String() == "true"
+}
+
+func stripHeader(b []byte) []byte {
+	if i := bytes.IndexByte(b, '\n'); i != -1 && bytes.HasPrefix(b, []byte("// Generated by Mochi")) {
+		return b[i+1:]
+	}
+	return b
+}
