+++ compiler/x/zig/tpcds_golden_test.go
@@ -4,7 +4,6 @@ package zigcode_test
 
 import (
 	"bytes"
-	"flag"
 	"fmt"
 	"os"
 	"os/exec"
@@ -17,20 +16,6 @@ import (
 	"mochi/runtime/vm"
 	"mochi/types"
 )
-
-var update = flag.Bool("update", false, "update golden files")
-
-func shouldUpdate() bool {
-	f := flag.Lookup("update")
-	return f != nil && f.Value.String() == "true"
-}
-
-func stripHeader(b []byte) []byte {
-	if i := bytes.IndexByte(b, '\n'); i != -1 && bytes.HasPrefix(b, []byte("// Generated by Mochi")) {
-		return b[i+1:]
-	}
-	return b
-}
 
 func TestZigCompiler_TPCDS_Golden(t *testing.T) {
 	zigc, err := zigcode.EnsureZig()
