+++ scripts/compile_rosetta_ocaml.go
@@ -1,3 +1,5 @@
+//go:build archive
+
 package main
 
 import (
