+++ tools/compilerust.go
@@ -1,4 +1,4 @@
-//go:build slow
+//go:build slow && compilerust
 
 package main
