+++ tools/rosetta/mochi_rust_golden_test.go
@@ -1,3 +1,5 @@
+//go:build slow
+
 package rosetta
 
 import (
